requirements.txt
flask>=3.0.0
werkzeug>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pymongo>=4.6.0
Pillow>=10.0.0
scikit-learn>=1.3.0
gunicorn>=21.2.0
# CPU-only torch — much smaller than full torch, faster to install on Render
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.0+cpu
torchvision==0.17.0+cpu